Dev replication js

+++ door-runtime/src/commonMain/kotlin/com/ustadmobile/door/entities/NodeIdAndAuth.kt
@@ -1,3 +1,3 @@
 package com.ustadmobile.door.entities
 
-data class NodeIdAndAuth (val nodeId: Int, val auth: String)
+data class NodeIdAndAuth (val nodeId: Long, val auth: String)
