+++ door-testdb/src/commonMain/kotlin/db2/ExampleEntityPkIntDao.kt
@@ -8,10 +8,10 @@ import androidx.room.Query
 abstract class ExampleEntityPkIntDao {
 
     @Insert
-    abstract fun insertAndReturnId(entity: ExampleEntityPkInt): Int
+    abstract fun insertAndReturnId(entity: ExampleEntityPkInt): Long
 
     @Insert
-    abstract fun insertListAndReturnIds(entityList: List<ExampleEntityPkInt>): List<Int>
+    abstract fun insertListAndReturnIds(entityList: List<ExampleEntityPkInt>): List<Long>
 
     @Query("SELECT * FROM ExampleEntityPkInt WHERE pk = :pk")
     abstract fun findByPk(pk: Int): ExampleEntityPkInt?
